+++ src/dolphin/workflows/stitch_and_unwrap.py
@@ -1,4 +1,5 @@
-from typing import List
+from pathlib import Path
+from typing import List, Tuple
 
 from dolphin import stitching, unwrap
 from dolphin._log import get_log, log_runtime
@@ -8,7 +9,11 @@ from .config import Workflow
 
 
 @log_runtime
-def run(ifg_list: List[VRTInterferogram], cfg: Workflow, debug: bool = False):
+def run(
+    ifg_list: List[VRTInterferogram],
+    cfg: Workflow,
+    debug: bool = False,
+) -> Tuple[List[Path], List[Path]]:
     """Run the displacement workflow on a stack of SLCs.
 
     Parameters
@@ -39,23 +44,37 @@ def run(ifg_list: List[VRTInterferogram], cfg: Workflow, debug: bool = False):
         file_date_fmt=cfg.inputs.cslc_date_fmt,
         output_dir=stitched_ifg_dir,
     )
-    # TODO: Stitch the correlation files
-    # tcorr_file = pl_path / "tcorr_average.tif"
+
+    # Stitch the correlation files
+    pl_path = cfg.phase_linking.directory
+    tcorr_files = list(pl_path.rglob("tcorr_average.tif"))
+    stitched_cor_file = stitched_ifg_dir / "tcorr_average.tif"
+    stitching.merge_images(
+        tcorr_files,
+        outfile=stitched_cor_file,
+        driver="GTiff",
+        overwrite=False,
+    )
 
     # #####################################
     # 2. Unwrap the stitched interferograms
     # #####################################
     if not cfg.unwrap_options.run_unwrap:
         logger.info("Skipping unwrap step")
-        return []
+        return [], []
 
     logger.info(f"Unwrapping interferograms in {stitched_ifg_dir}")
-    unwrapped_paths = unwrap.run(
+    # Compute the looks for the unwrapping
+    row_looks, col_looks = cfg.phase_linking.half_window.to_looks()
+    nlooks = row_looks * col_looks
+    unwrapped_paths, conncomp_paths = unwrap.run(
         ifg_path=stitched_ifg_dir,
         output_path=cfg.unwrap_options.directory,
-        cor_file=None,  # TODO: tcorr_file,
+        cor_file=stitched_cor_file,
+        nlooks=nlooks,
         # mask_file: Optional[Filename] = None,
-        max_jobs=20,
+        # TODO: max jobs based on the CPUs and the available RAM?
+        # max_jobs=20,
         # overwrite: bool = False,
         no_tile=True,
     )
@@ -65,4 +84,4 @@ def run(ifg_list: List[VRTInterferogram], cfg: Workflow, debug: bool = False):
     # ####################
     # TODO: Determine format for the tropospheric/ionospheric phase correction
 
-    return unwrapped_paths
+    return unwrapped_paths, conncomp_paths