+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Unreleased
 
+**Changed**
+
+- The `unwrap.py` module now uses isce3 to unwrap the interferogram.
+
+**Dependencies**
+
+
+Added requirements:
+
+- isce3>=0.8.0
+- pyproj>=3.3
+- Dropped support for Python 3.7
+
 # v0.0.4
 
 **Added**